Move Cas2 files from /api/model to /cas2/model.
