+++ src/main/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/api/model/UpdateCas2v2Application.kt
@@ -1,6 +1,7 @@
 package uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model
 
 import com.fasterxml.jackson.annotation.JsonProperty
+import uk.gov.justice.digital.hmpps.approvedpremisesapi.cas2.model.Cas2CohortDto
 import java.time.LocalDate
 
 data class UpdateCas2v2Application(
