+++ src/main/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/cas2/model/Cas2v2ApplicationStatusDetail.kt
@@ -1,16 +1,16 @@
-package uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model
+package uk.gov.justice.digital.hmpps.approvedpremisesapi.cas2.model
 
 import com.fasterxml.jackson.annotation.JsonProperty
 import io.swagger.v3.oas.annotations.media.Schema
 import java.util.UUID
 
-data class Cas2v2StatusUpdateDetail(
+data class Cas2v2ApplicationStatusDetail(
 
   @get:JsonProperty("id", required = true) val id: UUID,
 
-  @Schema(example = "moreInfoRequested", required = true, description = "")
+  @Schema(example = "changeOfCircumstances", required = true, description = "")
   @get:JsonProperty("name", required = true) val name: String,
 
-  @Schema(example = "More information requested", required = true, description = "")
+  @Schema(example = "Change of Circumstances", required = true, description = "")
   @get:JsonProperty("label", required = true) val label: String,
 )
