+++ src/main/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/cas2/model/Cas2v2StatusUpdateDetail.kt
@@ -1,16 +1,16 @@
-package uk.gov.justice.digital.hmpps.approvedpremisesapi.api.model
+package uk.gov.justice.digital.hmpps.approvedpremisesapi.cas2.model
 
 import com.fasterxml.jackson.annotation.JsonProperty
 import io.swagger.v3.oas.annotations.media.Schema
 import java.util.UUID
 
-data class Cas2v2ApplicationStatusDetail(
+data class Cas2v2StatusUpdateDetail(
 
   @get:JsonProperty("id", required = true) val id: UUID,
 
-  @Schema(example = "changeOfCircumstances", required = true, description = "")
+  @Schema(example = "moreInfoRequested", required = true, description = "")
   @get:JsonProperty("name", required = true) val name: String,
 
-  @Schema(example = "Change of Circumstances", required = true, description = "")
+  @Schema(example = "More information requested", required = true, description = "")
   @get:JsonProperty("label", required = true) val label: String,
 )
